+++ R/networks.R
@@ -1,14 +1,15 @@
 #' Get ASOS and AWOS networks
 #'
-#' @return a data.frame (tibble tibble) with the names and codes of available networks.
+#' @return a data.frame (tibble tibble) with the names and codes of
+#' available networks.
 #' @export
 #'
 #' @examples
 #' \dontrun{
 #' riem_networks()
 #' }
-riem_networks <- function(){
-  resp <- perform_riem_request(path = "api/1/networks.json")
+riem_networks <- function() {
+  resp <- perform_riem_request(path = "api/1/networks.json") # nolint: nonportable_path_linter
 
   httr2::resp_check_status(resp)
 
@@ -19,7 +20,7 @@ riem_networks <- function(){
     name = purrr::map_chr(content[["data"]], "name")
   )
 
-  is_asos_or_awos <- grepl("A[SW]OS", networks_data$code)
+  is_asos_or_awos <- grepl("A[SW]OS", networks_data[["code"]])
 
-  networks_data[is_asos_or_awos,]
+  networks_data[is_asos_or_awos, ]
 }