+++ NEWS.md
@@ -1,4 +1,10 @@
-# riem (development version)
+# riem 0.3.1
+
+* set tz=UTC on request to ensure tz result (#43, @akrherz)
+
+* Fixes timestamp parsing bug in riem_measures() caused by a lubridate 1.9.0 bug (#40, @BenoitFayolle)
+
+# riem 0.3.0
 
 * Switches to httr2 and httptest2 under the hood.
 * Improves error messages.